Allow standalone Analytics import I wanted to be able to stick all my analytics stuff onto a worker thread, but for iOS was unable to because several of the delegates were being duplicated. This setup is backwards compatible so that `analytics` can either be used off the main `firebase` module, or as a standalone module, with typings being correctly exported for both usages as well
The same thing could probably be done for some of the other portions of the plugin as well, like `firestore`, etc.
