+++ src/analytics/analytics.android.ts
@@ -1,0 +1,78 @@
+import * as appModule from "tns-core-modules/application";
+import { LogEventOptions, SetScreenNameOptions, SetUserPropertyOptions } from "./analytics";
+
+declare const com: any;
+
+export function logEvent(options: LogEventOptions): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      if (options.key === undefined) {
+        reject("Argument 'key' is missing");
+        return;
+      }
+
+      const bundle = new android.os.Bundle();
+      if (options.parameters !== undefined) {
+        for (const p in options.parameters) {
+          const param = options.parameters[p];
+          if (param.value !== undefined) {
+            bundle.putString(param.key, param.value);
+          }
+        }
+      }
+
+      com.google.firebase.analytics.FirebaseAnalytics.getInstance(
+        appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()
+      ).logEvent(options.key, bundle);
+
+      resolve();
+    } catch (ex) {
+      console.log("Error in firebase.analytics.logEvent: " + ex);
+      reject(ex);
+    }
+  });
+}
+
+export function setUserProperty(options: SetUserPropertyOptions): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      if (options.key === undefined) {
+        reject("Argument 'key' is missing");
+        return;
+      }
+      if (options.value === undefined) {
+        reject("Argument 'value' is missing");
+        return;
+      }
+
+      com.google.firebase.analytics.FirebaseAnalytics.getInstance(
+        appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()
+      ).setUserProperty(options.key, options.value);
+
+      resolve();
+    } catch (ex) {
+      console.log("Error in firebase.analytics.setUserProperty: " + ex);
+      reject(ex);
+    }
+  });
+}
+
+export function setScreenName(options: SetScreenNameOptions): Promise<any> {
+  return new Promise((resolve, reject) => {
+    try {
+      if (options.screenName === undefined) {
+        reject("Argument 'screenName' is missing");
+        return;
+      }
+
+      com.google.firebase.analytics.FirebaseAnalytics.getInstance(
+        appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()
+      ).setCurrentScreen(appModule.android.foregroundActivity, options.screenName, null);
+
+      resolve();
+    } catch (ex) {
+      console.log("Error in firebase.analytics.setScreenName: " + ex);
+      reject(ex);
+    }
+  });
+}
