+++ src/firebase.android.ts
@@ -522,74 +522,6 @@ firebase._isGooglePlayServicesAvailable = () => {
   return available;
 };
 
-firebase.analytics.logEvent = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.key === undefined) {
-        reject("Argument 'key' is missing");
-        return;
-      }
-
-      const bundle = new android.os.Bundle();
-      if (arg.parameters !== undefined) {
-        for (const p in arg.parameters) {
-          const param = arg.parameters[p];
-          if (param.value !== undefined) {
-            bundle.putString(param.key, param.value);
-          }
-        }
-      }
-
-      com.google.firebase.analytics.FirebaseAnalytics.getInstance(appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()).logEvent(arg.key, bundle);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.logEvent: " + ex);
-      reject(ex);
-    }
-  });
-};
-
-firebase.analytics.setUserProperty = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.key === undefined) {
-        reject("Argument 'key' is missing");
-        return;
-      }
-      if (arg.value === undefined) {
-        reject("Argument 'value' is missing");
-        return;
-      }
-
-      com.google.firebase.analytics.FirebaseAnalytics.getInstance(appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()).setUserProperty(arg.key, arg.value);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.setUserProperty: " + ex);
-      reject(ex);
-    }
-  });
-};
-
-firebase.analytics.setScreenName = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.screenName === undefined) {
-        reject("Argument 'screenName' is missing");
-        return;
-      }
-
-      com.google.firebase.analytics.FirebaseAnalytics.getInstance(appModule.android.currentContext || com.tns.NativeScriptApplication.getInstance()).setCurrentScreen(appModule.android.foregroundActivity, arg.screenName, null);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.setScreenName: " + ex);
-      reject(ex);
-    }
-  });
-};
-
 // see https://firebase.google.com/docs/admob/android/quick-start
 firebase.admob.showBanner = arg => {
   return new Promise((resolve, reject) => {
