+++ src/firebase.ios.ts
@@ -722,74 +722,6 @@ firebase.init = arg => {
   });
 };
 
-firebase.analytics.logEvent = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.key === undefined) {
-        reject("Argument 'key' is missing");
-        return;
-      }
-
-      const dic: any = NSMutableDictionary.new();
-      if (arg.parameters !== undefined) {
-        for (let p in arg.parameters) {
-          const param = arg.parameters[p];
-          if (param.value !== undefined) {
-            dic.setObjectForKey(param.value, param.key);
-          }
-        }
-      }
-
-      FIRAnalytics.logEventWithNameParameters(arg.key, dic);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.logEvent: " + ex);
-      reject(ex);
-    }
-  });
-};
-
-firebase.analytics.setUserProperty = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.key === undefined) {
-        reject("Argument 'key' is missing");
-        return;
-      }
-      if (arg.value === undefined) {
-        reject("Argument 'value' is missing");
-        return;
-      }
-
-      FIRAnalytics.setUserPropertyStringForName(arg.value, arg.key);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.setUserProperty: " + ex);
-      reject(ex);
-    }
-  });
-};
-
-firebase.analytics.setScreenName = arg => {
-  return new Promise((resolve, reject) => {
-    try {
-      if (arg.screenName === undefined) {
-        reject("Argument 'screenName' is missing");
-        return;
-      }
-
-      FIRAnalytics.setScreenNameScreenClass(arg.screenName, null);
-
-      resolve();
-    } catch (ex) {
-      console.log("Error in firebase.analytics.setScreenName: " + ex);
-      reject(ex);
-    }
-  });
-};
-
 firebase.admob.showBanner = arg => {
   return new Promise((resolve, reject) => {
     try {
